requirements-docker.txt
django>3.2
oauthlib>3.1
wheel
psycopg2-binary
django-formtools
django-ipware
django-reversion
django-recaptcha
docutils
pyjwt
cryptography
sorl-thumbnail
pillow
pytz
geopy
qrcode
requests
celery
python-u2flib-server
pymemcache
gunicorn
